refactor(dashboard-rpc): move the open-PR handoff decision out of the RPC layer
sendOpenPullRequest held the handoff business logic (refuse an empty or gone branch,
reuse an existing PR, derive the title/body) and prBodyFor beside it. That decision
belongs with the other handoff logic in run-handoff.ts, not in the RPC handler, which
should only resolve which run it is about. Add openSessionPullRequest there and reduce
the handler to resolve-then-delegate. Also fold the identical sendRemoveWorktree /
sendDeleteSession preamble (resolve path, refuse relay, stop preview before removal)
into one withWorktreeRemoval helper.
